+++ crates/sqlite-store/src/lib.rs
@@ -348,19 +348,12 @@ impl Store for SqliteStore {
         &self,
         account: &Account,
         initial_address: Address,
-        watch_only: bool,
     ) -> Result<(), StoreError> {
         let cloned_account = account.clone();
         let smt_forest = self.smt_forest.clone();
 
         self.interact_with_connection(move |conn| {
-            SqliteStore::insert_account(
-                conn,
-                &smt_forest,
-                &cloned_account,
-                &initial_address,
-                watch_only,
-            )
+            SqliteStore::insert_account(conn, &smt_forest, &cloned_account, &initial_address)
         })
         .await
     }
@@ -529,15 +522,9 @@ impl Store for SqliteStore {
         .await
     }
 
-    async fn remove_address(
-        &self,
-        address: Address,
-        account_id: AccountId,
-    ) -> Result<(), StoreError> {
-        self.interact_with_connection(move |conn| {
-            SqliteStore::remove_address(conn, &address, account_id)
-        })
-        .await
+    async fn remove_address(&self, address: Address) -> Result<(), StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::remove_address(conn, &address))
+            .await
     }
 
     async fn get_minimal_partial_account(